CellBackground moved to common components directory
