+++ mathesar_ui/src/components/CellBackground.svelte
