+++ mathesar_ui/src/components/cell-fabric/data-types/components/CellWrapper.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import CellBackground from '@mathesar/systems/table-view/row/CellBackground.svelte';
+  import CellBackground from '@mathesar/components/CellBackground.svelte';
   import { tick } from 'svelte';
   import type { HorizontalAlignment } from './typeDefinitions';
