+++ mathesar_ui/src/systems/table-view/row/GroupHeader.svelte
@@ -7,7 +7,7 @@
     ProcessedColumn,
   } from '@mathesar/stores/table-data';
   import type { RecordSummariesForSheet } from '@mathesar/stores/table-data/record-summaries/recordSummaryUtils';
-  import CellBackground from './CellBackground.svelte';
+  import CellBackground from '@mathesar/components/CellBackground.svelte';
   import GroupHeaderCellValue from './GroupHeaderCellValue.svelte';
 
   export let processedColumnsMap: Map<number, ProcessedColumn>;
